feat(cb2-6749): adds metadata to s3 upload to enable email via gov.notify, and refactors codebase
