+++ .devcontainer/postAttachCommand.sh
@@ -1,2 +1,3 @@
 #!/bin/bash
+
 git secrets --register-aws
