+++ src/handler.ts
@@ -1,13 +1,13 @@
 import { Handler, SQSBatchResponse, SQSEvent } from 'aws-lambda';
 import { TextDecoder } from 'util';
 import logger from './observability/logger';
-import { generateMinistryDocumentModel, generateTrlIntoServiceLetter } from './models/document';
+import { DocumentModel } from './models/documentModel';
 import { Request } from './models/request';
-import { DocumentType } from './models/documentName.enum';
 import { invokePdfGenLambda } from './services/Lamba.service';
 import { uploadPdfToS3 } from './services/S3.service';
+import { getDocumentFromRequest } from './models/documentModel.factory';
 
-const handler: Handler = async (event: SQSEvent): Promise<SQSBatchResponse> => {
+export const handler: Handler = async (event: SQSEvent): Promise<SQSBatchResponse> => {
   logger.debug("Function triggered'.");
   if (!event || !event.Records || !event.Records.length) {
     logger.error('ERROR: event is not defined.');
@@ -16,18 +16,9 @@ const handler: Handler = async (event: SQSEvent): Promise<SQSBatchResponse> => {
 
   const altPromiseArray = event.Records.map((sqsRecord) => {
     const request = JSON.parse(sqsRecord.body) as Request;
-    let documentData;
-    let fileName: string;
-    if (request.documentName === DocumentType.MINISTRY || request.documentName === DocumentType.MINISTRY_TRL) {
-      documentData = generateMinistryDocumentModel(request.vehicle, request.plate);
-      fileName = `plate_${request.plate.plateSerialNumber}`;
-    } else if (request.documentName === DocumentType.TRL_INTO_SERVICE) {
-      documentData = generateTrlIntoServiceLetter(request.vehicle, request.letter);
-      fileName = `letter_${request.vehicle.systemNumber}_${request.vehicle.vin}`;
-    } else {
-      throw new Error('Document Type not supported');
-    }
-    return generateAndUpload(documentData, request, fileName);
+
+    const document: DocumentModel = getDocumentFromRequest(request);
+    return generateAndUpload(document, request);
   });
 
   const results = await Promise.allSettled(altPromiseArray);
@@ -39,10 +30,10 @@ const handler: Handler = async (event: SQSEvent): Promise<SQSBatchResponse> => {
   };
 };
 
-const generateAndUpload = async (documentData, request: Request, fileName: string) => {
+const generateAndUpload = async (document: DocumentModel, request: Request) => {
   try {
-    logger.info(`Starting lambda to lambda invoke (data): ${JSON.stringify(documentData)}`);
-    const response = await invokePdfGenLambda(documentData, request.documentName);
+    logger.info(`Starting lambda to lambda invoke (data): ${JSON.stringify(document)}`);
+    const response = await invokePdfGenLambda(document, request.documentName);
     logger.info('Finished lambda to lambda invoke, checking response');
 
     if (response.StatusCode !== 200) {
@@ -59,21 +50,13 @@ const generateAndUpload = async (documentData, request: Request, fileName: strin
 
     // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
     const responseBuffer: Buffer = Buffer.from(responseJson.body, 'base64');
+    document.fileSize = responseBuffer.byteLength;
 
-    const metaData = {
-      'date-of-issue': Date.now().toString(),
-      'cert-type': request.documentName,
-      'file-format': 'pdf',
-      'file-size': responseBuffer.byteLength.toString(),
-      'should-email-certificate': 'false',
-    };
-    logger.info(`Starting s3 upload for file: ${process.env.BRANCH}/${fileName}`);
-    await uploadPdfToS3(responseBuffer, metaData, fileName);
+    logger.info(`Starting s3 upload for file: ${process.env.BRANCH}/${document.filename}`);
+    await uploadPdfToS3(responseBuffer, document.metaData, document.filename);
     logger.info('Finished s3 upload');
   } catch (error) {
     logger.error(error.message);
     throw error;
   }
 };
-
-export { handler, generateAndUpload };
