--- src/models/documentName.enum.ts
@@ -1,5 +1,0 @@
-export const enum DocumentType {
-  MINISTRY = 'VTG6_VTG7',
-  MINISTRY_TRL = 'VTG6_VTG7_TRL',
-  TRL_INTO_SERVICE = 'TrailerIntoService',
-}
