--- src/models/trlServiceLetter.ts
@@ -1,10 +1,0 @@
-import { IApplicantDetails } from './vehicleTechRecord';
-
-export type TrlServiceLetter = {
-  vin: string;
-  trailerId: string;
-  applicantDetails: IApplicantDetails;
-  letterDateRequested: string;
-  approvalTypeNumber: string;
-  paragraphId: number;
-};
